refactor(icons): move services from theme
