+++ src/framework/theme/components/icon/icon.module.ts
@@ -8,6 +8,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { NbIconComponent } from './icon.component';
+import { NbIconsLibrary } from './icons-library';
 
 @NgModule({
   imports: [
@@ -21,4 +22,10 @@ import { NbIconComponent } from './icon.component';
   ],
 })
 export class NbIconModule {
+
+  constructor(private iconsLibrary: NbIconsLibrary) {
+    // @breaking-change 4.0.0 remove and replace with eva-icons module
+    this.iconsLibrary.registerFontPack('nebular', { iconPrefix: 'nb' });
+    this.iconsLibrary.setDefaultPack('nebular');
+  }
 }
