+++ src/framework/theme/index.ts
@@ -99,6 +99,8 @@ export * from './components/tree-grid/data-source/tree-grid-data.service';
 export * from './components/tree-grid/data-source/tree-grid-filter.service';
 export * from './components/tree-grid/data-source/tree-grid.service';
 export * from './components/tree-grid/data-source/tree-grid-sort.service';
-export * from './services/icons';
 export * from './components/icon/icon.module';
 export * from './components/icon/icon.component';
+export * from './components/icon/icon';
+export * from './components/icon/icon-pack';
+export * from './components/icon/icons-library';
