--- src/framework/theme/services/icons/index.ts
@@ -1,3 +1,0 @@
-export * from './icon';
-export * from './icon-pack';
-export * from './icons-library';
